requirements.txt
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langgraph>=0.2.0
gradio>=4.44.0
langchain-core>=0.2.27
